Edit link fix
If page is not in draft or publish state extra check is done to see if the current user has been assigned the page via a workflow i.e. is an approver or publisher.  If they are then it is possible to access the page edit screen and make amends etc.